% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.R
\docType{data}
\name{citations}
\alias{citations}
\title{APBP database citation bibliographic identifiers}
\format{
A data frame with 157 rows and 3 variables:
\describe{
\item{citekey}{unique identifier for each citation}
\item{bibtype}{citation type, can be article (for journal articles), book, book chapter, conference paper, personal communication, report (for unpublished technical reports), or unpublished (for unpublished datasets, monographs, reports, supplements, theses, etc.). }
\item{year}{citation year}
}
}
\source{
\url{https://doi.org/10.3897/BDJ.11.e101476}
}
\usage{
citations
}
\description{
A dataset containing bibliographic identifiers for all citations
referenced in the APBP database.
}
\section{References}{

Che-Castaldo, C., G.R.W. Humphries, and H.J. Lynch. (2023). Antarctic Penguin Biogeography Project: Database of abundance and distribution for the Adélie, chinstrap, gentoo, emperor, macaroni and king penguin south of 60 S. \emph{Biodiversity Data Journal} 11: e101476, \url{https://doi.org/10.3897/BDJ.11.e101476}
}

\keyword{datasets}
