% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\docType{data}
\name{collections}
\alias{collections}
\title{Citation information for books or conference proceedings referenced in the APBP database}
\format{
A data frame with 12 rows and 11 variables:
\describe{
\item{collection_id}{unique identifier for each book or conference proceeding}
\item{citekey}{unique identifier for each citation}
\item{bibtype}{citation type, can be book or conference proceeding}
\item{title}{book or conference proceeding title}
\item{series}{book series, if applicable}
\item{volume}{series volume, if applicable}
\item{publisher}{book or conference proceeding publisher}
\item{city}{publisher city}
\item{country}{publisher country}
\item{isbn_issn}{ISBN or ISSN number, if available}
\item{url}{url link to book or conference proceeding, if available}
}
}
\source{
\url{https://doi.org/10.3897/BDJ.11.e101476}
}
\usage{
collections
}
\description{
A dataset containing citation information specific to books or conference proceedings
containing data included in the APBP database.
}
\section{References}{

Che-Castaldo, C., G.R.W. Humphries, and H.J. Lynch. (2023). Antarctic Penguin Biogeography Project: Database of abundance and distribution for the Adélie, chinstrap, gentoo, emperor, macaroni and king penguin south of 60 S. \emph{Biodiversity Data Journal} 11: e101476, \url{https://doi.org/10.3897/BDJ.11.e101476}
}

\keyword{datasets}
