% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapppd_bib.R
\docType{data}
\name{mapppd_bib}
\alias{mapppd_bib}
\title{Bibliography for citations referenced in the APBP database}
\format{
An object of class \code{BibEntry} (inherits from \code{bibentry}) of length 157.
}
\source{
\url{https://doi.org/10.3897/BDJ.11.e101476}
}
\usage{
mapppd_bib
}
\description{
Bibliographic information for all citations in the APBP database.
}
\section{References}{

Che-Castaldo, C., G.R.W. Humphries, and H.J. Lynch. (2023). Antarctic Penguin Biogeography Project: Database of abundance and distribution for the Adélie, chinstrap, gentoo, emperor, macaroni and king penguin south of 60 S. \emph{Biodiversity Data Journal} 11: e101476, url{https://doi.org/10.3897/BDJ.11.e101476}
}

\keyword{datasets}
