% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapppdr.R
\docType{package}
\name{mapppdr}
\alias{mapppdr}
\title{The 'mapppdr' package}
\description{
\code{mapppdr} is an R package used to access and visualize data from the Mapping Application for Penguin Populations and Projected Dynamics (MAPPPD) database.
}
\section{Details}{

The package contains 12 data frames, 1 \code{BibEntry} object, and 1 function:

\code{penguin_obs} - data frame containing all publicly available \emph{Pygoscelis} and Emperor penguin counts and presence-absence data included in MAPPPD at breeding sites south of 60 degrees S latitude

\code{citations} - data frame containing bibliographic identifiers for all citations referenced in MAPPPD

\code{sites} - data frame containing the name and location of all known \emph{Pygoscelis} and Emperor penguin breeding sites south of 60 degrees S latitude and all Antarctic Site Inventory (ASI) penguin/seabird census sites included in MAPPPD

\code{sites_sf} - \code{sites} data frame with locations stored as simple features POINT geometries

\code{species} - data frame containing the names of all penguin species included in MAPPPD

\code{site_species} - data frame linking \emph{Pygoscelis} and Emperor penguin species to their breeding sites in MAPPPD

\code{articles} - data frame containing citation information specific to journal articles whose data are included in MAPPPD

\code{collections} - data frame containing citation information specific to books or conference proceedings containing data included in MAPPPD

\code{incollections} - data frame containing citation information specific to book chapters or conference papers whose data are included in MAPPPD

\code{unpublished} - data frame containing citation information specific to unpublished data included in MAPPPD

\code{persons} - data frame containing bibliographic identifiers for all people who contributed data to MAPPPD

\code{citation_persons} - data frame linking contributors to their citations in MAPPPD

\code{mapppd_bib} - \code{BibEntry} object containing bibliographic information for all citations in MAPPPD

\code{penmap} - function that creates a leaflet map of all known \emph{Pygoscelis} and Emperor penguin breeding sites south of 60 degrees S latitude included in MAPPPD)

The Mapping Application for Penguin Populations and Projected Dynamics (MAPPPD) is a web-based, open access, decision-support tool designed to assist scientists, non-governmental organizations and policy-makers working to meet the management objectives as set forth by the Commission for the Conservation of Antarctic Marine Living Resources (CCAMLR) and other components of the Antarctic Treaty System (ATS). The MAPPPD database, described in Humphries et al. (2017), includes all publicly available published and unpublished count data on emperor, gentoo, Adélie and chinstrap penguins in Antarctica. A front end web interface located at \href{http://www.penguinmap.com}{www.penguinmap.com} provides free and ready access to count data alongside population models that infer abundance in missing data and allow for straightforward aggregation across multiple populations (see Che-Castaldo et al. 2017). The \code{mapppdr} package will always be updated before changes are reflected on the web interface at \href{http://www.penguinmap.com}{www.penguinmap.com} and should be considered the most up-to-date database available on the distribution and abundance of Antarctic penguins.

The vignette can be run using \code{vignette('mapppdr')} if vignette is built when installing package.
}

\section{Author(s)}{

Christian Che-Castaldo \href{mailto:christian.che-castaldo@stonybrook.edu}{christian.che-castaldo@stonybrook.edu}, Heather J. Lynch \href{mailto:heather.lynch@stonybrook.edu}{heather.lynch@stonybrook.edu}
}

\section{References}{

Humphries, G. R. W., R. Naveen, M. Schwaller, C. Che-Castaldo, P. McDowall, M. Schrimpf and H. J. Lynch. (2017). "Mapping Application for Penguin Populations and Projected Dynamics (MAPPPD): Data and tools for dynamic management and decision support". \emph{Polar Record} 53(2): 160-166,  \url{https://doi.org/10.1017/S0032247417000055}
}

