% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penguin_obs.R
\docType{data}
\name{penguin_obs}
\alias{penguin_obs}
\title{Count and presence/absence observations included in the APBP database}
\format{
A data frame with 4986 rows and 14 variables:
\describe{
\item{site_id}{unique identifier for each site}
\item{species_id}{unique identifier for each species}
\item{citekey}{unique identifier for each citation}
\item{month}{month of observation, if available}
\item{day}{day of observation, if available}
\item{doy}{day of year of observation, if available}
\item{date}{date of observation, if month and day are also known}
\item{year}{year of observation}
\item{season}{season of observation}
\item{type}{type of observation, either adults, nests (breeding pairs), or chicks}
\item{presence}{equals if species type (adults, nests (breeding pairs), or chicks) observed at the site and 0 otherwise}
\item{count}{count of species type (adults, nests (breeding pairs), or chicks) observed at the site, can be NA if record is presence/absence record only}
\item{accuracy}{accuracy scores were assigned on a 5 point scale (usually by the observer) that penguin census counters traditionally use to represent
count precision (see Croxall and Kirkwood, 1979)}
\item{vantage}{vantage point of observation, can be aerial (direct count using an aerial platform), aerial photo (count of a photograph taken from an aerial platform), ground (direct count by an individual on the ground), ground photo (count of a photograph taken from a position on the ground), landsat satellite, offshore vessel, sentinel satellite, uav (count of a photograph taken from an unmanned aerial vehicle), or vhr (population estimate based on very-high, < 5 m resolution, commercial satellite imagery)}
}
}
\source{
\url{https://doi.org/10.3897/BDJ.11.e101476}
}
\usage{
penguin_obs
}
\description{
A dataset containing the all publicly available \emph{Pygoscelis}, emperor, king and macaroni penguin
counts and presence-absence data at breeding sites south of 60 degrees S latitude contained within the APBP database.
}
\section{References}{

Che-Castaldo, C., G.R.W. Humphries, and H.J. Lynch. (2023). Antarctic Penguin Biogeography Project: Database of abundance and distribution for the Adélie, chinstrap, gentoo, emperor, macaroni and king penguin south of 60 S. \emph{Biodiversity Data Journal} 11: e101476, \url{https://doi.org/10.3897/BDJ.11.e101476}

Croxall, J.P. and E.D. Kirkwood. (1979). The distribution of penguins on the Antarctic Peninsula and islands of the Scotia Sea. British Antarctic Survey, Cambridge, United Kingdom
}

\keyword{datasets}
