% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penmap.R
\name{penmap}
\alias{penmap}
\title{Create leaflet of penguin breeding sites}
\usage{
penmap(species_map = c("ADPE", "GEPE", "CHPE", "EMPE", "KIPE", "MCPE"))
}
\arguments{
\item{species_map}{character vector of penguin species to include on map. Species are referenced using the species_id codes found in the species table.

Default \code{'c("ADPE", "GDPE", "CHPE", "EMPE", "KIPE", "MCPE")'} returns breeding sites for all penguin species in the APBP database.}
}
\description{
Creates leaflet of all known \emph{Pygoscelis}, emperor, king, and macaroni penguin breeding sites south of 60 degrees S latitude contained within the APBP database.
}
\section{Details}{

\code{species_id} argument must be a character vector.
}

\section{Notes}{

The map provides several helpful layers. These include the CCAMLR Statistical Areas and Divisions and Marine Protected Areas, the Antarctic Coastline (all exported from the \code{CCAMLRGIS} R package) and the Landsat Image Mosaic Of Antarctica (LIMA) (Bindschadler et al. 2008) WMS tiles hosted by the \href{https://www.pgc.umn.edu}{Polar Geospatial Center}.
}

\section{References}{

Bindschadler, R., P. Vornberger, A. Fleming, A. Fox, J. Mullins, D. Binnie, S. J. Paulsen, B. Granneman, and . Gorodetzky. (2008). "The Landsat image mosaic of Antarctica". \emph{Remote Sensing of Environment} 112(12): 4214-4226, \url{https://doi.org/10.1016/j.rse.2008.07.006}
Che-Castaldo, C., G.R.W. Humphries, and H.J. Lynch. (2023). Antarctic Penguin Biogeography Project: Database of abundance and distribution for the Adélie, chinstrap, gentoo, emperor, macaroni and king penguin south of 60 S. \emph{Biodiversity Data Journal} 11: e101476, \url{https://doi.org/10.3897/BDJ.11.e101476}
Stephane T., K. Reid and L. Robinson (2020). CCAMLRGIS: Antarctic Spatial Data Manipulation. R package version 3.0.6.  \url{https://CRAN.R-project.org/package=CCAMLRGIS}
}

\examples{
# Make default map
# penmap()

# Make map of just Adelie sites
# penmap("ADPE")

# Save leaflet in working directory
# This html file can be easily shared and opens in any web browser
# htmlwidgets::saveWidget(penmap("ADPE"), file = "penmap.html")

# Make map of adelie and chinstrap sites
# penmap(species_map = c("ADPE", "CHPE"))

}
