% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persons.R
\docType{data}
\name{persons}
\alias{persons}
\title{APBP database contributor bibliographic identifiers}
\format{
A data frame with 292 rows and 3 variables:
\describe{
\item{person_id}{unique identifier for each person}
\item{family}{family name of person}
\item{given}{given name of person}
}
}
\source{
\url{https://doi.org/10.3897/BDJ.11.e101476}
}
\usage{
persons
}
\description{
A dataset containing bibliographic identifiers for all references
contained within the APBP database.
}
\section{References}{

Che-Castaldo, C., G.R.W. Humphries, and H.J. Lynch. (2023). Antarctic Penguin Biogeography Project: Database of abundance and distribution for the Adélie, chinstrap, gentoo, emperor, macaroni and king penguin south of 60 S. \emph{Biodiversity Data Journal} 11: e101476, \url{https://doi.org/10.3897/BDJ.11.e101476}
}

\keyword{datasets}
