% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_species.R
\docType{data}
\name{site_species}
\alias{site_species}
\title{Name and species breeding at APBP database penguin breeding sites}
\format{
A data frame with 907 rows and 2 variables:
\describe{
\item{site_id}{unique identifier for each site}
\item{species_id}{unique identifier for each species}
}
}
\source{
\url{https://doi.org/10.3897/BDJ.11.e101476}
}
\usage{
site_species
}
\description{
A dataset containing the name of all known \emph{Pygoscelis} and emperor, king, and macaroni penguin breeding sites south of
60 degrees S latitude contained within the APBP database and which of these species breed there.
}
\section{References}{

Che-Castaldo, C., G.R.W. Humphries, and H.J. Lynch. (2023). Antarctic Penguin Biogeography Project: Database of abundance and distribution for the Adélie, chinstrap, gentoo, emperor, macaroni and king penguin south of 60 S. \emph{Biodiversity Data Journal} 11: e101476, \url{https://doi.org/10.3897/BDJ.11.e101476}
}

\keyword{datasets}
