% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\docType{data}
\name{species}
\alias{species}
\title{MAPPPD penguin species identifiers}
\format{
A data frame with 7 rows and 4 variables:
\describe{
\item{species_id}{unique identifier for each species}
\item{common_name}{species common name}
\item{genus}{genus}
\item{species}{species}
}
}
\source{
\url{http://www.penguinmap.com/}
}
\usage{
species
}
\description{
A dataset containing the names of all penguin species included in MAPPPD.
}
\section{References}{

Humphries, G. R. W., R. Naveen, M. Schwaller, C. Che-Castaldo, P. McDowall, M. Schrimpf and H. J. Lynch. (2017). "Mapping Application for Penguin Populations and Projected Dynamics (MAPPPD): Data and tools for dynamic management and decision support". \emph{Polar Record} 53(2): 160-166,  \url{https://doi.org/10.1017/S0032247417000055}
}

\keyword{datasets}
