% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpublished.R
\docType{data}
\name{unpublished}
\alias{unpublished}
\title{Citation information for persons referenced in MAPPPD}
\format{
A data frame with 10 rows and 4 variables:
\describe{
\item{citekey}{unique identifier for each citation}
\item{title}{description of data source}
\item{type}{type of unpublished data. These include things such as datasets, monographs, reports, supplements, thesis, etc.}
\item{note}{dadditional information, such as date downloaded, if available}
\item{doi}{dataset doi, if available}
\item{url}{url link to data source, if available}
}
}
\source{
\url{http://www.penguinmap.com/}
}
\usage{
unpublished
}
\description{
A dataset containing citation information specific to unpublished data included in MAPPPD.
}
\section{References}{

Humphries, G. R. W., R. Naveen, M. Schwaller, C. Che-Castaldo, P. McDowall, M. Schrimpf and H. J. Lynch. (2017). "Mapping Application for Penguin Populations and Projected Dynamics (MAPPPD): Data and tools for dynamic management and decision support". \emph{Polar Record} 53(2): 160-166,  \url{https://doi.org/10.1017/S0032247417000055}
}

\keyword{datasets}
