% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.R
\docType{data}
\name{articles}
\alias{articles}
\title{Citation information for journal articles referenced in MAPPPD}
\format{
A data frame with 106 rows and 10 variables:
\describe{
\item{citekey}{unique identifier for each citation}
\item{title}{article title}
\item{journal}{journal name}
\item{volume}{journal volume}
\item{number}{journal number, if available}
\item{start_page}{starting page of article}
\item{end_page}{ending page of article}
\item{doi}{article doi, if available}
\item{url}{url link to article, if available}
\item{early_access}{boolean indicator that is TRUE if the article is currently early access and FALSE otherwise}

}
}
\source{
\url{http://www.penguinmap.com/}
}
\usage{
articles
}
\description{
A dataset containing citation information specific to journal articles
whose data are included in MAPPPD.
}
\section{References}{

Humphries, G. R. W., R. Naveen, M. Schwaller, C. Che-Castaldo, P. McDowall, M. Schrimpf and H. J. Lynch. (2017). "Mapping Application for Penguin Populations and Projected Dynamics (MAPPPD): Data and tools for dynamic management and decision support". \emph{Polar Record} 53(2): 160-166, \url{https://doi.org/10.1017/S0032247417000055}
}

\keyword{datasets}
