% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_persons.R
\docType{data}
\name{citation_persons}
\alias{citation_persons}
\title{Linking contributors to their citations referenced in MAPPPD}
\format{
A data frame with 588 rows and 4 variables:
\describe{
\item{citekey}{unique identifier for each citation}
\item{person_id}{unique identifier for each person}
\item{role}{identifies whether person was an author or editor}
\item{cite_order}{identifies the author or editor number}
}
}
\source{
\url{http://www.penguinmap.com/}
}
\usage{
citation_persons
}
\description{
A dataset linking contributors to their citations in MAPPPD.
}
\section{References}{

Humphries, G. R. W., R. Naveen, M. Schwaller, C. Che-Castaldo, P. McDowall, M. Schrimpf and H. J. Lynch. (2017). "Mapping Application for Penguin Populations and Projected Dynamics (MAPPPD): Data and tools for dynamic management and decision support". \emph{Polar Record} 53(2): 160-166, https://doi.org/10.1017/S0032247417000055
}

\keyword{datasets}
