% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.R
\docType{data}
\name{citations}
\alias{citations}
\title{MAPPPD citation bibliographic identifiers}
\format{
A data frame with 187 rows and 3 variables:
\describe{
\item{citekey}{unique identifier for each citation}
\item{bibtype}{citation type}
\item{year}{citation year}
}
}
\source{
\url{http://www.penguinmap.com/}
}
\usage{
citations
}
\description{
A dataset containing bibliographic identifiers for all citations
referenced in MAPPPD.
}
\section{References}{

Humphries, G. R. W., R. Naveen, M. Schwaller, C. Che-Castaldo, P. McDowall, M. Schrimpf and H. J. Lynch. (2017). "Mapping Application for Penguin Populations and Projected Dynamics (MAPPPD): Data and tools for dynamic management and decision support". \emph{Polar Record} 53(2): 160-166, https://doi.org/10.1017/S0032247417000055
}

\keyword{datasets}
