% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapppd_bib.R
\docType{data}
\name{mapppd_bib}
\alias{mapppd_bib}
\title{Bibliography for citations referenced in MAPPPD}
\format{
An object of class \code{BibEntry} (inherits from \code{bibentry}) of length 194.
}
\source{
\url{http://www.penguinmap.com/}
}
\usage{
mapppd_bib
}
\description{
Bibliographic information for all citations in MAPPPD.
}
\section{References}{

Humphries, G. R. W., R. Naveen, M. Schwaller, C. Che-Castaldo, P. McDowall, M. Schrimpf and H. J. Lynch. (2017). "Mapping Application for Penguin Populations and Projected Dynamics (MAPPPD): Data and tools for dynamic management and decision support". \emph{Polar Record} 53(2): 160-166,  \url{https://doi.org/10.1017/S0032247417000055}
}

\keyword{datasets}
