% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persons.R
\docType{data}
\name{persons}
\alias{persons}
\title{MAPPPD contributor bibliographic identifiers}
\format{
A data frame with 325 rows and 3 variables:
\describe{
\item{person_id}{unique identifier for each person}
\item{family}{family name of person}
\item{given}{given name of person}
}
}
\source{
\url{http://www.penguinmap.com/}
}
\usage{
persons
}
\description{
A dataset containing bibliographic identifiers for all references
contained within MAPPPD.
}
\section{References}{

Humphries, G. R. W., R. Naveen, M. Schwaller, C. Che-Castaldo, P. McDowall, M. Schrimpf and H. J. Lynch. (2017). "Mapping Application for Penguin Populations and Projected Dynamics (MAPPPD): Data and tools for dynamic management and decision support". \emph{Polar Record} 53(2): 160-166,  \url{https://doi.org/10.1017/S0032247417000055}
}

\keyword{datasets}
