% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_species.R
\docType{data}
\name{site_species}
\alias{site_species}
\title{Name and species breeding at MAPPPD penguin breeding sites}
\format{
A data frame with 829 rows and 2 variables:
\describe{
\item{site_id}{unique identifier for each site}
\item{species_id}{unique identifier for each species}
}
}
\source{
\url{http://www.penguinmap.com/}

\url{https://oceanites.org/research-portal/antarctic-site-inventory/asi-census-sites/}
}
\usage{
site_species
}
\description{
A dataset containing the name of all known Pygoscelis and Emperor penguin breeding sites south of
60 degrees S latitude contained within MAPPPD and which of these species breed there.
}
\section{References}{

Humphries, G. R. W., R. Naveen, M. Schwaller, C. Che-Castaldo, P. McDowall, M. Schrimpf and H. J. Lynch. (2017). "Mapping Application for Penguin Populations and Projected Dynamics (MAPPPD): Data and tools for dynamic management and decision support". \emph{Polar Record} 53(2): 160-166,  \url{https://doi.org/10.1017/S0032247417000055}
}

\keyword{datasets}
