% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites_sf.R
\docType{data}
\name{sites_sf}
\alias{sites_sf}
\title{Name and location of APBP database breeding sites}
\format{
An sf data frame with 720 rows, 4 attribute variables, and 1 simple features geometry column:
\describe{
\item{site_id}{unique identifier for each site}
\item{site_name}{site name and/or description}
\item{region}{site region as reported in the 2015 ASI Antarctic Peninsula Compendium (Naveen and Lynch, 2015)}
\item{ccamlr_id}{Commission for the Conservation of Antarctic Marine Living Resources (CCAMLR) subarea within which the site lies}
\item{geometry}{simple features POINT geometries of all site location (EPSG:4326)}
}
}
\source{
\url{https://doi.org/10.3897/BDJ.11.e101476}
}
\usage{
sites_sf
}
\description{
A dataset containing the name and location of all known \emph{Pygoscelis}, emperor,
king and macaroni penguin breeding sites south of 60 degrees S latitude
contained within the APBP database.
}
\section{References}{

Che-Castaldo, C., G.R.W. Humphries, and H.J. Lynch. (2023). Antarctic Penguin Biogeography Project: Database of abundance and distribution for the Adélie, chinstrap, gentoo, emperor, macaroni and king penguin south of 60 S. \emph{Biodiversity Data Journal} 11: e101476, \url{https://doi.org/10.3897/BDJ.11.e101476}

Naveen, R. and H.J. Lynch. (2015). "Antarctic Peninsula Compendium" 3rd Edition, Chevy Chase, Maryland, USA.
}

\keyword{datasets}
