% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\docType{data}
\name{species}
\alias{species}
\title{APBP database penguin species identifiers}
\format{
A data frame with 7 rows and 4 variables:
\describe{
\item{species_id}{unique identifier for each species}
\item{common_name}{species common name}
\item{genus}{genus}
\item{species}{species}
}
}
\source{
\url{https://doi.org/10.3897/BDJ.11.e101476}
}
\usage{
species
}
\description{
A dataset containing the names of all penguin species included in the APBP database.
}
\section{References}{

Che-Castaldo, C., G.R.W. Humphries, and H.J. Lynch. (2023). Antarctic Penguin Biogeography Project: Database of abundance and distribution for the Adélie, chinstrap, gentoo, emperor, macaroni and king penguin south of 60 S. \emph{Biodiversity Data Journal} 11: e101476, \url{https://doi.org/10.3897/BDJ.11.e101476}
}

\keyword{datasets}
