% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpublished.R
\docType{data}
\name{unpublished}
\alias{unpublished}
\title{Citation information for persons referenced in the APBP database}
\format{
A data frame with 11 rows and 4 variables:
\describe{
\item{citekey}{unique identifier for each citation}
\item{title}{description of data source}
\item{type}{type of unpublished data. These include things such as datasets, monographs, reports, supplements, thesis, etc.}
\item{note}{dadditional information, such as date downloaded, if available}
\item{doi}{dataset doi, if available}
\item{url}{url link to data source, if available}
}
}
\source{
\url{https://doi.org/10.3897/BDJ.11.e101476}
}
\usage{
unpublished
}
\description{
A dataset containing citation information specific to unpublished data included in the APBP database.
}
\section{References}{

Che-Castaldo, C., G.R.W. Humphries, and H.J. Lynch. (2023). Antarctic Penguin Biogeography Project: Database of abundance and distribution for the Adélie, chinstrap, gentoo, emperor, macaroni and king penguin south of 60 S. \emph{Biodiversity Data Journal} 11: e101476, \url{https://doi.org/10.3897/BDJ.11.e101476}
}

\keyword{datasets}
